notes resolved
