+++ package.json
@@ -81,7 +81,7 @@
     "react": "17.0.2",
     "react-native": "0.66.3",
     "react-native-device-country": "1.0.2",
-    "react-native-device-info": "^10.0.2",
+    "react-native-device-info": "10.0.2",
     "react-native-fs": "2.18.0",
     "react-native-keyboard-aware-scroll-view": "0.9.5",
     "react-native-modal-datetime-picker": "13.0.0",
