+++ src/api/ApiClient.ts
@@ -4,16 +4,12 @@ import _filter from 'lodash/filter'
 import _get from 'lodash/get'
 import _map from 'lodash/map'
 import _sortBy from 'lodash/sortBy'
-import {
-  getBrand,
-  getDeviceType,
-  getProduct,
-  getSystemName,
-} from 'react-native-device-info'
+import { getBrand, getProduct } from 'react-native-device-info'
 
 import { IOnCheckoutSuccess } from '..'
 import { IWaitingListFields } from '../components/waitingList/types'
 import { Config } from '../helpers/Config'
+import { switchGetDeviceType, switchGetSystemName } from '../helpers/Device'
 import {
   deleteAllData,
   getData,
@@ -77,30 +73,6 @@ axiosRetry(Client, { retries: 3 })
 Client.interceptors.request.use(async (config: AxiosRequestConfig) => {
   const guestToken = await getData(LocalStorageKeys.AUTH_GUEST_TOKEN)
   const accessToken = await getData(LocalStorageKeys.ACCESS_TOKEN)
-
-  const switchGetDeviceType = () => {
-    const deviceType = getDeviceType()
-    switch (deviceType) {
-      case 'Tablet':
-        return 'Tablet'
-      case 'Handset':
-        return 'Mobile Device'
-      case 'Desktop':
-        return 'Desktop'
-      default:
-        return 'unknown'
-    }
-  }
-
-  const switchGetSystemName = () => {
-    const deviceType = getSystemName()
-    switch (deviceType) {
-      case 'Android':
-        return 'Android OS'
-      default:
-        return 'iPod, iPhone & iPad'
-    }
-  }
 
   if (config.headers) {
     config.headers['Device-Info'] = JSON.stringify({
